==> mpranalyze_analysis/1_MPRAnalyze_Prepare_Annotation_Data.ipynb <==
# MPRAnalyze Prepare Annotation Data

library(tidyverse)

# annotations for DNA data
dna_annot <- data.frame(version = factor(rep(c("V1", "V2"), each=150)), batch = factor(rep(c(21:26), each=50)), 
                        barcode = factor(rep(c(1:50), times=3))) %>% 
                        unite("label", version, batch, barcode, sep=":", remove=FALSE) %>% 
                        column_to_rownames("label")

# annotations for RNA data
rna_annot_pool <- data.frame(version = factor(rep(c("V1", "V2"), each=500)), batch = factor(rep(c(1:20), each=50)), 
                        barcode = factor(rep(c(1:50), times=10))) %>% 
                        unite("label", version, batch, barcode, sep=":", remove=FALSE) %>% 
                        column_to_rownames("label")

# save DNA annotations
write.table(dna_annot, '../data/dna_annot.txt')

# save RNA annotations
write.table(rna_annot_pool, '../data/rna_annot_pool.txt', sep = "\t")

# get size of DNA and RNA counts
size_dna <- 600
size_rna <- 2000

# make DNA annotations for allelic comparison
dna_annot_allelic <- data.frame(version = factor(rep(c("V1", "V2"), each=size_dna/2)), 
                                batch = factor(rep(c(11:13), each=100)), 
                                barcode = factor(rep(c(1:50), each=2, times=(size_dna/100))), 
                                alleletype = factor(rep(c('ref', 'alt'), times=(size_dna/2)))) %>% 
                        unite("label", version, batch, barcode, alleletype, sep=":", remove=FALSE) %>% 
                        column_to_rownames("label")

# find interaction between barcodes and allele types
dna_annot_allelic$barcode_allelic <- interaction(dna_annot_allelic$barcode, dna_annot_allelic$alleletype)
dna_annot_allelic$barcode_allelic_version <- interaction(dna_annot_allelic$version, dna_annot_allelic$barcode_allelic)

# make RNA annotations for allelic comparison
rna_annot_allelic_pool <- data.frame(version = factor(rep(c("V1", "V2"), each=size_rna/2)), 
                                batch = factor(rep(c(1:10), each=100)), 
                                barcode = factor(rep(c(1:50), each=2, times=(size_rna/100))), 
                                alleletype = factor(rep(c('ref', 'alt'), times=(size_rna/2)))) %>% 
                        unite("label", version, batch, barcode, alleletype, sep=":", remove=FALSE) %>% 
                        column_to_rownames("label")

# find interaction between barcodes, allele types, and versions
rna_annot_allelic_pool$barcode_allelic <- interaction(rna_annot_allelic_pool$barcode, rna_annot_allelic_pool$alleletype)
rna_annot_allelic_pool$barcode_allelic_version <- interaction(rna_annot_allelic_pool$version, rna_annot_allelic_pool$barcode_allelic)

# save DNA allelic annotations
write.table(dna_annot_allelic, '../data/merged_dna_annot_allelic.txt')

# save RNA allelic annotations
write.table(rna_annot_allelic_pool, '../data/merged_rna_annot_pool_allelic.txt')
