==> mpranalyze_analysis/3a_MPRAnalyze_Allelic_Cmp_Prepare_Count_Data.ipynb <==
# MPRAnalyze Workflow for Preparing Count Data for Allelic Comparison

library(tidyverse)
library(ggplot2)

## Pool1

# read in dataset
mpra_v1 <- read_tsv("../data/mpra_qtigc_pgl4v1_pool1.txt")
mpra_v2 <- read_tsv("../data/mpra_qtigc_pgl4v2_pool1.txt")

num_enhancers_v1 <- dim(mpra_v1)[1]
num_enhancers_v2 <- dim(mpra_v2)[1]

# format DNA counts for v1
dna_counts_v1 <- mpra_v1 %>% select(-starts_with("rna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v1/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("dna")) %>%
    rename_with(~paste0("V1.", .), -testcre)

# format DNA counts for v2
dna_counts_v2 <- mpra_v2 %>% select(-starts_with("rna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v2/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("dna")) %>%
    rename_with(~paste0("V2.", .), -testcre)

# merge v1 and v2 DNA counts
dna_counts <- dna_counts_v1 %>% inner_join(dna_counts_v2, by = "testcre") %>%
    column_to_rownames("testcre") %>% as.matrix()

head(dna_counts)

# format RNA counts
rna_counts_v1 <- mpra_v1 %>% select(-starts_with("dna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v1/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("rna")) %>%
    rename_with(~paste0("V1.", .), -testcre)

# format RNA counts
rna_counts_v2 <- mpra_v2 %>% select(-starts_with("dna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v2/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("rna")) %>%
    rename_with(~paste0("V2.", .), -testcre)

# merge v1 and v2 RNA counts
rna_counts <- rna_counts_v1 %>% inner_join(rna_counts_v2, by = "testcre") %>%
    column_to_rownames("testcre") %>% as.matrix()

# split row names from DNA counts into enhancer information
dna_enc_names <- data.frame(do.call(rbind, strsplit(rownames(dna_counts), "_")))
colnames(dna_enc_names) <- c("chrom", "pos", "snp", "allele", "mutation", "alleletype")

# expand dna counts to have ref and alt counts on same rows for each enhancer
dna_counts_allelic <- as.data.frame(dna_counts) %>% 
    mutate(dna_enc_names) %>% 
    pivot_wider(names_from=alleletype, values_from=contains('dna')) %>% 
    unite(enhancer, chrom, pos, snp, allele, mutation, sep='_') %>%
    column_to_rownames('enhancer')

# split row names from RNA counts into enhancer information
rna_enc_names <- data.frame(do.call(rbind, strsplit(rownames(rna_counts), "_")))
colnames(rna_enc_names) <- c("chrom", "pos", "snp", "allele", "mutation", "alleletype")

# expand rna counts to have ref and alt counts on same rows for each enhancer
rna_counts_allelic <- as.data.frame(rna_counts) %>% 
    mutate(rna_enc_names) %>% 
    pivot_wider(names_from=alleletype, values_from=contains('rna')) %>% 
    unite(enhancer, chrom, pos, snp, allele, mutation, sep='_') %>%
    column_to_rownames('enhancer')

# save DNA allelic counts
write.table(dna_counts_allelic, '../data/pool1_dna_counts_allelic.txt')

# save RNA allelic counts
write.table(rna_counts_allelic, '../data/pool1_rna_counts_allelic.txt')

## Pool2

# read in dataset
mpra_v1 <- read_tsv("../data/mpra_qtigc_pgl4v1_pool2.txt")
mpra_v2 <- read_tsv("../data/mpra_qtigc_pgl4v2_pool2.txt")

num_enhancers_v1 <- dim(mpra_v1)[1]
num_enhancers_v2 <- dim(mpra_v2)[1]

# format DNA counts for v1
dna_counts_v1 <- mpra_v1 %>% select(-starts_with("rna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v1/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("dna")) %>%
    rename_with(~paste0("V1.", .), -testcre)

# format DNA counts for v2
dna_counts_v2 <- mpra_v2 %>% select(-starts_with("rna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v2/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("dna")) %>%
    rename_with(~paste0("V2.", .), -testcre)

# merge v1 and v2 DNA counts
dna_counts <- dna_counts_v1 %>% inner_join(dna_counts_v2, by = "testcre") %>%
    column_to_rownames("testcre") %>% as.matrix()

head(dna_counts)

# format RNA counts
rna_counts_v1 <- mpra_v1 %>% select(-starts_with("dna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v1/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("rna")) %>%
    rename_with(~paste0("V1.", .), -testcre)

# format RNA counts
rna_counts_v2 <- mpra_v2 %>% select(-starts_with("dna"), -barcode, -snp) %>% 
    mutate(row=rep(c(1:50), times=num_enhancers_v2/50)) %>% 
    pivot_wider(names_from=row, values_from=starts_with("rna")) %>%
    rename_with(~paste0("V2.", .), -testcre)

# remove low quality replicate (v1:r2 and v2:r7)
rna_counts_v1 <- rna_counts_v1 %>%
    select(-c(starts_with("V1.rna.r2")))

rna_counts_v2 <- rna_counts_v2 %>%
    select(-c(starts_with("V2.rna.r7")))

# merge v1 and v2 RNA counts
rna_counts <- rna_counts_v1 %>% inner_join(rna_counts_v2, by = "testcre") %>%
    column_to_rownames("testcre") %>% as.matrix()

# split row names from DNA counts into enhancer information
dna_enc_names <- data.frame(do.call(rbind, strsplit(rownames(dna_counts), "_")))
colnames(dna_enc_names) <- c("chrom", "pos", "snp", "allele", "mutation", "alleletype")

# expand dna counts to have ref and alt counts on same rows for each enhancer
dna_counts_allelic <- as.data.frame(dna_counts) %>% 
    mutate(dna_enc_names) %>% 
    pivot_wider(names_from=alleletype, values_from=contains('dna')) %>% 
    unite(enhancer, chrom, pos, snp, allele, mutation, sep='_') %>%
    column_to_rownames('enhancer')

# split row names from RNA counts into enhancer information
rna_enc_names <- data.frame(do.call(rbind, strsplit(rownames(rna_counts), "_")))
colnames(rna_enc_names) <- c("chrom", "pos", "snp", "allele", "mutation", "alleletype")

# expand rna counts to have ref and alt counts on same rows for each enhancer
rna_counts_allelic <- as.data.frame(rna_counts) %>% 
    mutate(rna_enc_names) %>% 
    pivot_wider(names_from=alleletype, values_from=contains('rna')) %>% 
    unite(enhancer, chrom, pos, snp, allele, mutation, sep='_') %>%
    column_to_rownames('enhancer')

# save DNA allelic counts
write.table(dna_counts_allelic, '../data/pool2_dna_counts_allelic.txt')

# save RNA allelic counts
write.table(rna_counts_allelic, '../data/pool2_rna_counts_allelic.txt')

---

Note: we need to modify RNA annotation data for pool 2

rna_annot_allelic_p2 <- read.table('../data/merged_rna_annot_pool_allelic.txt')

head(rna_annot_allelic_p2)

rna_annot_allelic_p2 <- rna_annot_allelic_p2 %>%
    mutate_if(rep(TRUE, ncol(rna_annot_allelic_p2)), as.factor)

head(rna_annot_allelic_p2)

# adjust rna_annot accordingly
rna_annot_allelic_p2 <- rna_annot_allelic_p2 %>%
    filter(!(version=="V1" & batch == 2)) %>%
    filter(!(version=="V2" & batch == 7))

write.table(rna_annot_allelic_p2, '../data/merged_rna_annot_pool_allelic_p2.txt')
