==> mpranalyze_analysis/4_MPRAnalyze_Merge_Allelic_Comparison_Results.ipynb <==
# MPRAnalyze Merge Allelic Comparison Results

library(tidyverse)

## Pool1

# find all merged files for pool 1
list_results <- list.files(path = "../results/pool1", recursive = FALSE,
                            pattern = "pool1_mpranalyze_allelic_results_[0-9]", 
                            full.names = TRUE)

# check file names
list_results

# check that number of merged files matches number of array tasks
length(list_results)

# read contents of all merged files into list
allelic_results_df <- lapply(list_results, function(x) {read.table(file = x, 
                                                                   header = T, 
                                                                   sep=" ")})

# combine all rows into one dataframe
merged_r_allelic <- bind_rows(allelic_results_df, .id = "column_label") %>% select(-c("column_label"))

# check dimensions of combined results
dim(merged_r_allelic)

# add row indices as row names
rownames(merged_r_allelic)<-c(1:dim(merged_r_allelic)[1])

head(merged_r_allelic)

colnames(merged_r_allelic)

# save merged results as text file
write.table(merged_r_allelic, "../results/pool1_mpranalyze_allelic_results_merged.txt")

## Pool2

# find all merged files for pool 2
list_results <- list.files(path = "../results/pool2", recursive = FALSE,
                            pattern = "pool2_mpranalyze_allelic_results_[0-9]", 
                            full.names = TRUE)

# check file names
list_results

# check that number of merged files matches number of array tasks
length(list_results)

# read contents of all merged files into list
allelic_results_df <- lapply(list_results, function(x) {read.table(file = x, 
                                                                   header = T, 
                                                                   sep=" ")})

# combine all rows into one dataframe
merged_r_allelic <- bind_rows(allelic_results_df, .id = "column_label") %>% select(-c("column_label"))

# check dimensions of combined results
dim(merged_r_allelic)

# add row indices as row names
rownames(merged_r_allelic)<-c(1:dim(merged_r_allelic)[1])

head(merged_r_allelic)

colnames(merged_r_allelic)

# save merged results as text file
write.table(merged_r_allelic, "../results/pool2_mpranalyze_allelic_results_merged.txt")
