==> mpranalyze_analysis/5_Calculate_CPM.ipynb <==
# MPRA Calculate and Filter for CPM

library(tidyverse)

## Pool1

# import DNA counts for Pool 1
dna_counts_pool <- read.table("../data/pool1_dna_counts_allelic.txt", header=TRUE)

head(dna_counts_pool)

# aggregate DNA counts
master_pool <- dna_counts_pool %>%  
    mutate(V1_DNA_ref_count = rowSums(select(., starts_with("V1") & ends_with("ref")))) %>% 
    mutate(V1_DNA_alt_count = rowSums(select(., starts_with("V1") & ends_with("alt")))) %>% 
    mutate(V2_DNA_ref_count = rowSums(select(., starts_with("V2") & ends_with("ref")))) %>% 
    mutate(V2_DNA_alt_count = rowSums(select(., starts_with("V2") & ends_with("alt")))) %>% 
    mutate(V1_DNA_ref_CPM = (V1_DNA_ref_count / (sum(V1_DNA_ref_count) + sum(V1_DNA_alt_count))) * 1000000) %>%
    mutate(V1_DNA_alt_CPM = (V1_DNA_alt_count / (sum(V1_DNA_ref_count) + sum(V1_DNA_alt_count))) * 1000000) %>%
    mutate(V2_DNA_ref_CPM = (V2_DNA_ref_count / (sum(V2_DNA_ref_count) + sum(V2_DNA_alt_count))) * 1000000) %>%
    mutate(V2_DNA_alt_CPM = (V2_DNA_alt_count / (sum(V2_DNA_ref_count) + sum(V2_DNA_alt_count))) * 1000000) %>%
    mutate(DNA_CPM_QC = ifelse((V1_DNA_ref_CPM>=8 & V1_DNA_alt_CPM>=8) | 
                                 (V2_DNA_ref_CPM>=8 & V2_DNA_alt_CPM>=8), TRUE, FALSE)) %>%
    select(V1_DNA_ref_count, V1_DNA_alt_count, V2_DNA_ref_count, V2_DNA_alt_count, V1_DNA_ref_CPM, 
           V1_DNA_alt_CPM, V2_DNA_ref_CPM, V2_DNA_alt_CPM, DNA_CPM_QC) %>% 
           rownames_to_column("testcre")

head(master_pool)

# split DNA count row names into enhancer information
pool_names <- data.frame(do.call(rbind, strsplit(master_pool$testcre, "_")))
colnames(pool_names) <- c("chrom", "hg19_pos", "snp", "allele", "mutation")

# add enhancer information to master table
master_pool <- cbind(pool_names, master_pool)

head(master_pool)

# make position 1-based
master_pool <- master_pool %>% mutate(hg19_pos = as.numeric(hg19_pos) + 1)

head(master_pool)

# save counts and CPM values for Pool
write.table(master_pool, "../results/pool1_dna_counts_cpm.txt")

## Pool2

# import DNA counts for Pool 1
dna_counts_pool <- read.table("../data/pool2_dna_counts_allelic.txt", header=TRUE)

head(dna_counts_pool)

# aggregate DNA counts
master_pool <- dna_counts_pool %>%  
    mutate(V1_DNA_ref_count = rowSums(select(., starts_with("V1") & ends_with("ref")))) %>% 
    mutate(V1_DNA_alt_count = rowSums(select(., starts_with("V1") & ends_with("alt")))) %>% 
    mutate(V2_DNA_ref_count = rowSums(select(., starts_with("V2") & ends_with("ref")))) %>% 
    mutate(V2_DNA_alt_count = rowSums(select(., starts_with("V2") & ends_with("alt")))) %>% 
    mutate(V1_DNA_ref_CPM = (V1_DNA_ref_count / (sum(V1_DNA_ref_count) + sum(V1_DNA_alt_count))) * 1000000) %>%
    mutate(V1_DNA_alt_CPM = (V1_DNA_alt_count / (sum(V1_DNA_ref_count) + sum(V1_DNA_alt_count))) * 1000000) %>%
    mutate(V2_DNA_ref_CPM = (V2_DNA_ref_count / (sum(V2_DNA_ref_count) + sum(V2_DNA_alt_count))) * 1000000) %>%
    mutate(V2_DNA_alt_CPM = (V2_DNA_alt_count / (sum(V2_DNA_ref_count) + sum(V2_DNA_alt_count))) * 1000000) %>%
    mutate(DNA_CPM_QC = ifelse((V1_DNA_ref_CPM>=8 & V1_DNA_alt_CPM>=8) | 
                                 (V2_DNA_ref_CPM>=8 & V2_DNA_alt_CPM>=8), TRUE, FALSE)) %>%
    select(V1_DNA_ref_count, V1_DNA_alt_count, V2_DNA_ref_count, V2_DNA_alt_count, V1_DNA_ref_CPM, 
           V1_DNA_alt_CPM, V2_DNA_ref_CPM, V2_DNA_alt_CPM, DNA_CPM_QC) %>% 
           rownames_to_column("testcre")

head(master_pool)

# split DNA count row names into enhancer information
pool_names <- data.frame(do.call(rbind, strsplit(master_pool$testcre, "_")))
colnames(pool_names) <- c("chrom", "hg19_pos", "snp", "allele", "mutation")

# add enhancer information to master table
master_pool <- cbind(pool_names, master_pool)

head(master_pool)

# make position 1-based
master_pool <- master_pool %>% mutate(hg19_pos = as.numeric(hg19_pos) + 1)

head(master_pool)

# save counts and CPM values for Pool
write.table(master_pool, "../results/pool2_dna_counts_cpm.txt")
