==> mpranalyze_analysis/6_MPRA_Enhancer_Allelic_Significance.ipynb <==
# MPRA Find Significant Activity on All Versions and Pools

library(MPRAnalyze)
library(tidyverse)
library(ggplot2)

## Pool 1

# load MPRAObjects
pool_alpha <- read.table("../results/pool1_mpranalyze_results.txt")

# load DNA CPM counts
pool_cpm <- read.table("../results/pool1_dna_counts_cpm.txt")

# load results from allelic significance
pool_allelic <- read.table("../results/pool1_mpranalyze_allelic_results_merged.txt", header=TRUE)

head(pool_alpha)

head(pool_cpm)

head(pool_allelic)

head(pool_alpha)

# log-transform alpha values
pool_alpha <- pool_alpha %>% mutate(log2a = log2(alpha))

# calculate z-score with alpha
pool_alpha <- pool_alpha %>% mutate(zscore = (log2a - mean(log2a)) / sd(log2a))

head(pool_alpha)

# widen table to separate alleletypes and find enhancer significance
pool_alpha_w <- pool_alpha %>% 
    pivot_wider(names_from=type, values_from=c(alpha, log2a, zscore))

# select relevant columns from DNA CPM counts
pool_cpm <- pool_cpm %>% select(snp, DNA_CPM_QC)

head(pool_cpm)

# add DNA CPM classifications to v1 pool1 table
pool_alpha_w <- pool_alpha_w %>% left_join(pool_cpm, by = c("snp"))

head(pool_alpha_w)

# classify variants that are enhancers
pool_enhancer <- pool_alpha_w %>% filter(DNA_CPM_QC) %>% mutate(is_enhancer=(zscore_ref > 0 | zscore_alt > 0))

# join enhancer classifications with overall table
pool <- pool_alpha_w %>% left_join(pool_enhancer)

head(pool)

# extract allelic p-values
pool <- pool %>% left_join(select(pool_allelic, c(snp, logFC, lrt_pval)), by = "snp") %>%
    rename(pval_allelic = lrt_pval)

head(pool)

# calculate FDR for only enhancers
pool_enc <- pool %>% filter(is_enhancer == TRUE) %>% 
    mutate(fdr_allelic=p.adjust(pval_allelic, method = "BH")) %>% select(snp, fdr_allelic)

# add FDR values to table
pool <- pool %>% left_join(pool_enc, by = 'snp')

head(pool)

# find number of elements that are significant with FDR <= 0.05
sum(pool$fdr_allelic<=0.01, na.rm=TRUE)

# find allelic significance
pool_all <- pool %>% filter(is_enhancer) %>% mutate(is_allelic = (fdr_allelic <= 0.01) & (abs(logFC)>=log(1.1)))
pool1 <- pool %>% left_join(select(pool_all, c(snp, is_allelic)), by = "snp")

# save enhancer and allelic significance results
write.table(pool1, "../results/pool1_final_result.txt", 
            sep="\t", quote = FALSE, row.names = FALSE, col.names = TRUE)

## Pool 2

# load MPRAObjects
pool_alpha <- read.table("../results/pool2_mpranalyze_results.txt")

# load DNA CPM counts
pool_cpm <- read.table("../results/pool2_dna_counts_cpm.txt")

# load results from allelic significance
pool_allelic <- read.table("../results/pool2_mpranalyze_allelic_results_merged.txt", header=TRUE)

head(pool_alpha)

head(pool_cpm)

head(pool_allelic)

head(pool_alpha)

# log-transform alpha values
pool_alpha <- pool_alpha %>% mutate(log2a = log2(alpha))

# calculate z-score with alpha
pool_alpha <- pool_alpha %>% mutate(zscore = (log2a - mean(log2a)) / sd(log2a))

head(pool_alpha)

# widen table to separate alleletypes and find enhancer significance
pool_alpha_w <- pool_alpha %>% 
    pivot_wider(names_from=type, values_from=c(alpha, log2a, zscore))

# select relevant columns from DNA CPM counts
pool_cpm <- pool_cpm %>% select(snp, DNA_CPM_QC)

head(pool_cpm)

# add DNA CPM classifications to v1 pool1 table
pool_alpha_w <- pool_alpha_w %>% left_join(pool_cpm, by = c("snp"))

head(pool_alpha_w)

# classify variants that are enhancers
pool_enhancer <- pool_alpha_w %>% filter(DNA_CPM_QC) %>% mutate(is_enhancer=(zscore_ref > 0 | zscore_alt > 0))

# join enhancer classifications with overall table
pool <- pool_alpha_w %>% left_join(pool_enhancer)

head(pool)

# extract allelic p-values
pool <- pool %>% left_join(select(pool_allelic, c(snp, logFC, lrt_pval)), by = "snp") %>%
    rename(pval_allelic = lrt_pval)

head(pool)

# calculate FDR for only enhancers
pool_enc <- pool %>% filter(is_enhancer == TRUE) %>% 
    mutate(fdr_allelic=p.adjust(pval_allelic, method = "BH")) %>% select(snp, fdr_allelic)

# add FDR values to table
pool <- pool %>% left_join(pool_enc, by = 'snp')

head(pool)

# find number of elements that are significant with FDR <= 0.05
sum(pool$fdr_allelic<=0.01, na.rm=TRUE)

# find allelic significance
pool_all <- pool %>% filter(is_enhancer) %>% mutate(is_allelic = (fdr_allelic <= 0.01) & (abs(logFC)>=log(1.1)))
pool2 <- pool %>% left_join(select(pool_all, c(snp, is_allelic)), by = "snp")

head(pool2)

# save enhancer and allelic significance results
write.table(pool2, "../results/pool2_final_result.txt", 
            sep="\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
